requirements.txt
# Python Dependencies for Document Digitization Tool
# Python 3.8+ required

# OpenAI API client for OCR and table processing
openai>=1.0.0

# Flask web framework
flask>=2.0.0
python-dotenv>=0.19.0
flask-cors>=4.0.0

# Flask extensions (optional, if needed)
flask-jsonpify
flask-sqlalchemy
flask-restful

# Database drivers
mysql-connector-python>=8.0.0
pymysql>=1.0.0

# Swagger/OpenAPI documentation
flasgger>=0.9.7.1

# Cloudflare R2 (S3-compatible) storage
boto3>=1.26.0

# HTTP requests for API calls
requests>=2.28.0

# Note: Pandoc is also required but must be installed separately via system package manager
# See README.md for installation instructions:
# - Windows: choco install pandoc or winget install --id JohnMacFarlane.Pandoc
# - macOS: brew install pandoc
# - Linux: sudo apt-get install pandoc (Ubuntu/Debian) or sudo dnf install pandoc (Fedora)